extract info_fields into its own module
