+++ src/onefetch/license.rs
@@ -2,6 +2,8 @@ use askalono::{Store, TextData};
 
 use crate::onefetch::error::*;
 
+pub const LICENSE_FILES: [&str; 3] = ["LICENSE", "LICENCE", "COPYING"];
+
 static CACHE_DATA: &[u8] = include_bytes!("../../resources/licenses/cache.bin.zstd");
 const MIN_THRESHOLD: f32 = 0.8;
