+++ src/onefetch/mod.rs
@@ -4,6 +4,7 @@ pub mod commit_info;
 pub mod error;
 pub mod image_backends;
 pub mod info;
+pub mod info_fields;
 pub mod language;
 pub mod license;
 pub mod options;
